refactor(myaccount): reorganize account & profile screens into subpackages
Move MyAccount and UserProfile screen/viewmodel files into internal/myaccount
and internal/userprofile subpackages and update imports/tests accordingly.
